🐛 fix(runtime): 修复 ESP 串口帧长度编译
将 VLVT 长度两个字节的组合结果显式转换为 uint16_t，避免 ESP-IDF 的列表初始化在整型提升后触发 -Werror=narrowing。\n\n验证：CLANG_FORMAT=clang-format RUFF=ruff ./scripts/check_format.sh、serial_voice_protocol_test、git diff --check 通过；本机 ESP-IDF 历史构建目录无 CMake cache，目标编译由 CI 复核。\n\nRefs #296

+++ components/voicelife_runtime/src/serial_voice_test.cc
@@ -120,7 +120,8 @@ class SerialVoiceTest::Impl final {
             const detail::SerialVoiceFrameHeader frame_header{
                 .version = header[0],
                 .kind = header[1],
-                .payload_bytes = static_cast<uint16_t>(header[2]) | (static_cast<uint16_t>(header[3]) << 8U),
+                .payload_bytes =
+                    static_cast<uint16_t>(static_cast<uint16_t>(header[2]) | (static_cast<uint16_t>(header[3]) << 8U)),
             };
             if (!detail::IsValidSerialVoiceHeader(frame_header)) {
                 ESP_LOGW(kTag, "SERIAL_VOICE_FRAME_REJECT version=%u kind=%u length=%u",
